requirements.txt
streamlit==1.41.1

google-generativeai>=0.8.0

gTTS==2.5.4

numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
supabase
